% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_import_data.R
\name{covid19_import_data_action}
\alias{covid19_import_data_action}
\title{covid19_import_data (action)}
\usage{
covid19_import_data_action(data, argset, tables)
}
\arguments{
\item{data}{Data}

\item{argset}{Argset}

\item{tables}{DB tables}
}
\description{
covid19_import_data (action)
}
