% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_import_data.R
\name{covid19_import_data_data_selector}
\alias{covid19_import_data_data_selector}
\title{covid19_import_data (data selector)}
\usage{
covid19_import_data_data_selector(argset, tables)
}
\arguments{
\item{argset}{Argset}

\item{tables}{DB tables}
}
\description{
covid19_import_data (data selector)
}
