% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/household_incomes_and_house_prices_fit_model_and_find_outliers.R
\name{household_incomes_and_house_prices_fit_model_and_find_outliers_action}
\alias{household_incomes_and_house_prices_fit_model_and_find_outliers_action}
\title{household_incomes_and_house_prices_fit_model_and_find_outliers (action)}
\usage{
household_incomes_and_house_prices_fit_model_and_find_outliers_action(
  data,
  argset,
  schema
)
}
\arguments{
\item{data}{Data}

\item{argset}{Argset}

\item{schema}{DB Schema}
}
\description{
household_incomes_and_house_prices_fit_model_and_find_outliers (action)
}
