% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/household_incomes_and_house_prices_import_data.R
\name{household_incomes_and_house_prices_import_data_action}
\alias{household_incomes_and_house_prices_import_data_action}
\title{household_incomes_and_house_prices_import_data (action)}
\usage{
household_incomes_and_house_prices_import_data_action(data, argset, schema)
}
\arguments{
\item{data}{Data}

\item{argset}{Argset}

\item{schema}{DB Schema}
}
\description{
household_incomes_and_house_prices_import_data (action)
}
