% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/household_incomes_and_house_prices_import_data.R
\name{household_incomes_and_house_prices_import_data_data_selector}
\alias{household_incomes_and_house_prices_import_data_data_selector}
\title{household_incomes_and_house_prices_import_data (data selector)}
\usage{
household_incomes_and_house_prices_import_data_data_selector(argset, schema)
}
\arguments{
\item{argset}{Argset}

\item{schema}{DB Schema}
}
\description{
household_incomes_and_house_prices_import_data (data selector)
}
