% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_env_and_namespace.R
\docType{data}
\name{global}
\alias{global}
\title{Declaration of environments that can be used globally}
\format{
An object of class \code{environment} of length 2.
}
\usage{
global
}
\description{
Declaration of environments that can be used globally
}
\keyword{datasets}
