% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_clean_data.R
\name{weather_clean_data_data_selector}
\alias{weather_clean_data_data_selector}
\title{weather_clean_data (data selector)}
\usage{
weather_clean_data_data_selector(argset, tables)
}
\arguments{
\item{argset}{Argset}

\item{tables}{DB tables}
}
\description{
weather_clean_data (data selector)
}
