% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_download_and_import_rawdata.R
\name{weather_download_and_import_rawdata_action}
\alias{weather_download_and_import_rawdata_action}
\title{weather_download_and_import_rawdata (action)}
\usage{
weather_download_and_import_rawdata_action(data, argset, tables)
}
\arguments{
\item{data}{Data}

\item{argset}{Argset}

\item{tables}{DB tables}
}
\description{
weather_download_and_import_rawdata (action)
}
