% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_download_and_import_rawdata.R
\name{weather_download_and_import_rawdata_data_selector}
\alias{weather_download_and_import_rawdata_data_selector}
\title{weather_download_and_import_rawdata (data selector)}
\usage{
weather_download_and_import_rawdata_data_selector(argset, tables)
}
\arguments{
\item{argset}{Argset}

\item{tables}{DB tables}
}
\description{
weather_download_and_import_rawdata (data selector)
}
