% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_export_plots.R
\name{weather_export_plots_action}
\alias{weather_export_plots_action}
\title{weather_export_plots (action)}
\usage{
weather_export_plots_action(data, argset, tables)
}
\arguments{
\item{data}{Data}

\item{argset}{Argset}

\item{tables}{DB tables}
}
\description{
weather_export_plots (action)
}
