% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_export_plots.R
\name{weather_export_plots_data_selector}
\alias{weather_export_plots_data_selector}
\title{weather_export_plots (data selector)}
\usage{
weather_export_plots_data_selector(argset, tables)
}
\arguments{
\item{argset}{Argset}

\item{tables}{DB tables}
}
\description{
weather_export_plots (data selector)
}
